refactor: 서비스 인터페이스 제거
