+++ src/test/java/com/codeit/team4/deokhugam/review/service/ReviewServiceTest.java
@@ -34,10 +34,10 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.dao.DataIntegrityViolationException;
 
 @ExtendWith(MockitoExtension.class)
-class ReviewServiceImplTest {
+class ReviewServiceTest {
 
     @InjectMocks
-    private ReviewServiceImpl reviewService;
+    private ReviewService reviewService;
 
     @Mock
     private ReviewRepository reviewRepository;
